refactor: Separate ClustersTable class
