+++ src/widgets/components/__init__.py
@@ -1,2 +1,3 @@
 from .clustering_steps import ClusteringStepsTemplate
+from .clustering_table import ClustersTable
 from .parameters import ParametersGroupBox
