+++ src/widgets/results_widgets/gmm_results.py
@@ -3,26 +3,20 @@ from functools import partial
 import numpy as np
 import pandas as pd
 from matplotlib import pyplot as plt
-from PyQt5.QtCore import Qt
 from PyQt5.QtWidgets import (
     QComboBox,
     QFormLayout,
     QGroupBox,
     QHBoxLayout,
-    QInputDialog,
     QLabel,
-    QMessageBox,
     QPushButton,
     QSpinBox,
-    QTableView,
     QVBoxLayout,
-    QWidget,
 )
 
 from algorithms import get_samples
 from visualization import ClusteringCanvas
-from widgets import QtTable
-from widgets.components import ParametersGroupBox
+from widgets.components import ClustersTable, ParametersGroupBox
 from widgets.results_widgets import AlgorithmResultsWidget
 
 
@@ -35,7 +29,6 @@ class GMMResultsWidget(AlgorithmResultsWidget):
         self.max_label = np.amax(self.labels) + 1
         self.mean = pd.DataFrame(mean, columns=self.columns)
         self.sigma = sigma
-        self.selected_cluster = None
         self.layout = QHBoxLayout(self)
 
         self.num_samples = min(35, self.data.shape[0] // 2)
@@ -84,35 +77,25 @@ class GMMResultsWidget(AlgorithmResultsWidget):
         self.results_canvas = ClusteringCanvas(self.fig, axes, False)
         self.clustering_group_layout.addWidget(self.results_canvas, 1)
 
+        self.layout.addWidget(self.clustering_result_group, 1)
+
         # cluster details
-        self.clusters_details_group = QGroupBox()
-        self.clusters_group_layout = QVBoxLayout(self.clusters_details_group)
-        self.clusters_details_group.setTitle("Cluster")
-        self.clusters_table = QTableView()
-        self.clusters_table.setModel(QtTable(self.mean.round(3)))
-        self.clusters_table.doubleClicked.connect(self.show_cluster)
-        for i in range(len(self.columns)):
-            self.clusters_table.setColumnWidth(i, 120)
-        self.clusters_table_header = QWidget()
-        self.clusters_table_header_layout = QHBoxLayout()
-        self.clusters_table_instruction = QLabel(
-            "Double click on any field to preview a cluster"
+        self.clusters_group = QGroupBox()
+        self.clusters_group_layout = QVBoxLayout(self.clusters_group)
+        self.clusters_group.setTitle("Clusters")
+
+        self.clusters_table = ClustersTable(
+            self.data, self.labels, self.mean, len(self.columns)
         )
-        self.save_all_button = QPushButton("SAVE RESULTS")
-        self.save_all_button.clicked.connect(
-            partial(self.on_save_button_click, self.data.assign(cluster=self.labels))
-        )
-        self.save_all_button.setFixedWidth(120)
-        self.clusters_table_header_layout.addWidget(self.clusters_table_instruction)
-        self.clusters_table_header_layout.addWidget(self.save_all_button)
-        self.clusters_table_header.setLayout(self.clusters_table_header_layout)
+        self.clusters_table.table_changed.connect(self.update_cluster_plot)
+        self.clusters_group_layout.addWidget(self.clusters_table, 1)
+
         self.fig_distributions, ax = plt.subplots(1, 1)
         self.clusters_canvas = ClusteringCanvas(self.fig_distributions, ax, False)
-        self.clusters_group_layout.addWidget(self.clusters_table_header)
-        self.clusters_group_layout.addWidget(self.clusters_table, 1)
+
         self.clusters_group_layout.addWidget(self.clusters_canvas, 1)
-        self.layout.addWidget(self.clustering_result_group, 1)
-        self.layout.addWidget(self.clusters_details_group, 1)
+        self.layout.addWidget(self.clusters_group, 1)
+
         self.update_plot()
         self.update_cluster_plot()
 
@@ -159,11 +142,11 @@ class GMMResultsWidget(AlgorithmResultsWidget):
         )
 
     def update_cluster_plot(self):
-        if self.selected_cluster is not None:
+        if self.clusters_table.selected_cluster is not None:
             indexes = [
                 i
                 for i in range(len(self.labels))
-                if self.labels[i] == self.selected_cluster
+                if self.labels[i] == self.clusters_table.selected_cluster
             ]
             x = self.data.iloc[indexes][self.ox]
             y = self.data.iloc[indexes][self.oy]
@@ -180,10 +163,10 @@ class GMMResultsWidget(AlgorithmResultsWidget):
                 self.columns.get_loc(self.oy),
             ]
             mean = [
-                x_means.iloc[self.selected_cluster],
-                y_means.iloc[self.selected_cluster],
+                x_means.iloc[self.clusters_table.selected_cluster],
+                y_means.iloc[self.clusters_table.selected_cluster],
             ]
-            sigma_helper = self.sigma[self.selected_cluster]
+            sigma_helper = self.sigma[self.clusters_table.selected_cluster]
             sigma = [
                 [sigma_helper[x_index][x_index], sigma_helper[x_index][y_index]],
                 [sigma_helper[y_index][x_index], sigma_helper[y_index][y_index]],
@@ -194,7 +177,7 @@ class GMMResultsWidget(AlgorithmResultsWidget):
                 y,
                 mean,
                 sigma,
-                self.selected_cluster,
+                self.clusters_table.selected_cluster,
                 len(x_means),
                 self.ox,
                 self.oy,
@@ -245,69 +228,3 @@ class GMMResultsWidget(AlgorithmResultsWidget):
                 min_y - sep_y,
                 max_y + sep_y,
             )
-
-    def show_cluster(self):
-        self.selected_cluster = (
-            self.clusters_table.selectionModel().selectedIndexes()[0].row()
-        )
-        rows = [
-            i
-            for i in range(len(self.labels))
-            if self.labels[i] == self.selected_cluster
-        ]
-        elements = self.data.iloc[rows]
-        self.clusters_table.setModel(QtTable(elements))
-        buttons_widget = QWidget()
-        buttons_layout = QHBoxLayout()
-        exit_button = QPushButton("X")
-        exit_button.clicked.connect(self.exit_from_cluster)
-        exit_button.setFixedWidth(50)
-        save_button = QPushButton("SAVE")
-        save_button.clicked.connect(partial(self.on_save_button_click, elements))
-        save_button.setFixedWidth(100)
-        buttons_layout.addWidget(save_button)
-        buttons_layout.addWidget(exit_button)
-        buttons_layout.setAlignment(Qt.AlignRight)
-        buttons_widget.setLayout(buttons_layout)
-        self.clusters_group_layout.insertWidget(0, buttons_widget)
-        self.clusters_table_header.hide()
-        self.clusters_table.doubleClicked.disconnect()
-        self.update_cluster_plot()
-
-    def exit_from_cluster(self):
-        self.clusters_table.setModel(QtTable(self.mean.round(3)))
-        self.clusters_group_layout.itemAt(0).widget().setParent(None)
-        self.clusters_table_header.show()
-        self.clusters_table.doubleClicked.connect(self.show_cluster)
-        self.selected_cluster = None
-        self.update_cluster_plot()
-
-    def on_save_button_click(self, elements):
-        path, is_ok = QInputDialog.getText(self, "Save to file", "Enter filename")
-        if is_ok and path:
-            if not path.endswith(".csv"):
-                path += ".csv"
-            try:
-                elements.to_csv(path)
-            except:
-                error = QMessageBox()
-                error.setIcon(QMessageBox.Critical)
-                error.setText(
-                    "Something wrong happened while writing data to file. Try again."
-                )
-                error.setWindowTitle("Saving failed")
-                error.exec_()
-        elif not is_ok:
-            pass
-        elif not path:
-            error = QMessageBox()
-            error.setIcon(QMessageBox.Critical)
-            error.setText("No path was provided")
-            error.setWindowTitle("Empty path")
-            error.exec_()
-        else:
-            error = QMessageBox()
-            error.setIcon(QMessageBox.Critical)
-            error.setText("This file extension is not supported.")
-            error.setWindowTitle("Unsupported extension")
-            error.exec_()
